Zoom Rooms: review suggestions, add missing handler
