+++ zoom_rooms/indico_zoom_rooms/templates/explanation.html
@@ -2,6 +2,6 @@
     href="{{ docs_url }}"
     style="margin: 0;"
     class="ui basic mini blue button"
-    title="{% trans %}Room &quot;{{room_name}}&quot; is equipped with a Zoom Room. Click to find out more.{% endtrans %}">
+    title="{% trans %}Room &quot;{{ room_name }}&quot; is equipped with a Zoom Room. Click to find out more.{% endtrans %}">
     <i class="icon check circle"></i> Zoom Rooms
 </a>
